Move entity alliance system from AbilityRegistry to EntityManager
The allied/attack check system (EntityAlliedCheck, AlliedAttackCheck,
EntityPetCheck) is entity-relationship logic, not ability logic. Move
the check interfaces from ability/check/ to entity/check/ and the
registration + query methods from AbilityRegistry to EntityManager.
Resolves the TODO that was already flagged in AbilityRegistry. Guardian
tasks now use EntityManager.areEntitiesAllied() and also hoist the
caster resolution before the entity loop to avoid redundant lookups.
