+++ src/main/java/us/eunoians/mcrpg/entity/EntityManager.java
@@ -1,9 +1,14 @@
 package us.eunoians.mcrpg.entity;
 
+import com.diamonddagger590.mccore.pair.ImmutablePair;
+import com.diamonddagger590.mccore.pair.Pair;
 import com.diamonddagger590.mccore.registry.manager.Manager;
-import org.bukkit.entity.Player;
+import org.bukkit.NamespacedKey;
+import org.bukkit.entity.Entity;
 import org.jetbrains.annotations.NotNull;
 import us.eunoians.mcrpg.McRPG;
+import us.eunoians.mcrpg.entity.check.AlliedAttackCheck;
+import us.eunoians.mcrpg.entity.check.EntityAlliedCheck;
 import us.eunoians.mcrpg.entity.holder.AbilityHolder;
 import us.eunoians.mcrpg.entity.holder.QuestHolder;
 
@@ -13,7 +18,8 @@ import java.util.Optional;
 import java.util.UUID;
 
 /**
- * Tracks all entities that are different kinds of holders
+ * Tracks all entities that are different kinds of holders and manages entity relationship
+ * checks such as alliance and attack eligibility.
  * <p>
  * {@link AbilityHolder AbilityHolders} stored in this tracker may be more
  * than just an {@link AbilityHolder}, such as an {@link us.eunoians.mcrpg.entity.holder.SkillHolder} in the
@@ -23,11 +29,15 @@ public class EntityManager extends Manager<McRPG> {
 
     private final Map<UUID, AbilityHolder> abilityHolderMap;
     private final Map<UUID, QuestHolder> questHolderMap;
+    private final Map<NamespacedKey, EntityAlliedCheck> entityAlliedFunctions;
+    private final Map<NamespacedKey, AlliedAttackCheck> alliedAttackCheckFunctions;
 
     public EntityManager(@NotNull McRPG mcRPG) {
         super(mcRPG);
         this.abilityHolderMap = new HashMap<>();
         this.questHolderMap = new HashMap<>();
+        this.entityAlliedFunctions = new HashMap<>();
+        this.alliedAttackCheckFunctions = new HashMap<>();
     }
 
     /**
@@ -110,15 +120,88 @@ public class EntityManager extends Manager<McRPG> {
     }
 
     /**
-     * Checks if the target player is allied with the entity identified by the given UUID.
-     * Allied entities should not be affected by hostile ability effects (knockback, slowness, pull, etc.).
+     * Register the provided {@link EntityAlliedCheck} to be checked when {@link #areEntitiesAllied(Entity, Entity)} is called.
+     * <p>
+     * This also registers {@link AlliedAttackCheck#DEFAULT_ALLIED_ATTACK_CHECK_FUNCTION} as the default function to prevent allies
+     * from attacking each other. This behavior can be overridden by calling {@link #registerAlliedAttackCheckFunction(NamespacedKey, AlliedAttackCheck)} with
+     * a different implementation.
      *
-     * @param casterUUID The UUID of the entity that cast the ability.
-     * @param target     The target player to check alliance with.
-     * @return {@code true} if the target is allied with the caster, {@code false} otherwise.
+     * @param namespacedKey        The {@link NamespacedKey} to register this {@link EntityAlliedCheck} against.
+     * @param entityAlliedFunction The {@link EntityAlliedCheck} to register.
      */
-    public boolean isAllied(@NotNull UUID casterUUID, @NotNull Player target) {
-        // TODO: Implement with party system
+    public void registerEntityAlliedFunction(@NotNull NamespacedKey namespacedKey, @NotNull EntityAlliedCheck entityAlliedFunction) {
+        entityAlliedFunctions.put(namespacedKey, entityAlliedFunction);
+        registerAlliedAttackCheckFunction(namespacedKey, AlliedAttackCheck.DEFAULT_ALLIED_ATTACK_CHECK_FUNCTION);
+    }
+
+    /**
+     * Checks to see if the two provided {@link Entity entities} are allies or not using registered {@link EntityAlliedCheck EntityAlliedFunctions}.
+     * <p>
+     * This allows 3rd party plugins to anonymously register handling for their specific definition of what an "ally" is.
+     * <p>
+     * The order of the two entities should not matter as well.
+     *
+     * @param entity1       The first {@link Entity} to check.
+     * @param entity2       The second {@link Entity} to check.
+     * @param namespacedKey The {@link NamespacedKey} of the allied function to check.
+     * @return {@code true} if the two {@link Entity entities} are considered allies by the registered {@link EntityAlliedCheck}.
+     */
+    public boolean areEntitiesAllied(@NotNull Entity entity1, @NotNull Entity entity2, @NotNull NamespacedKey namespacedKey) {
+        if (entityAlliedFunctions.containsKey(namespacedKey)) {
+            return entityAlliedFunctions.get(namespacedKey).areAllies(entity1, entity2);
+        }
         return false;
+    }
+
+    /**
+     * Checks to see if the two provided {@link Entity entities} are allies or not using registered {@link EntityAlliedCheck EntityAlliedFunctions}.
+     * <p>
+     * This allows 3rd party plugins to anonymously register handling for their specific definition of what an "ally" is.
+     * <p>
+     * The order of the two entities should not matter as well.
+     *
+     * @param entity1 The first {@link Entity} to check.
+     * @param entity2 The second {@link Entity} to check.
+     * @return {@code true} if the two {@link Entity entities} are considered allies by any registered {@link EntityAlliedCheck EntityAlliedFunctions}.
+     */
+    public Pair<Boolean, Optional<NamespacedKey>> areEntitiesAllied(@NotNull Entity entity1, @NotNull Entity entity2) {
+        for (NamespacedKey namespacedKey : entityAlliedFunctions.keySet()) {
+            if (areEntitiesAllied(entity1, entity2, namespacedKey)) {
+                return ImmutablePair.of(true, Optional.of(namespacedKey));
+            }
+        }
+        return ImmutablePair.of(false, Optional.empty());
+    }
+
+    /**
+     * Register the provided {@link AlliedAttackCheck} to be checked when {@link #areEntitiesAllied(Entity, Entity)} is called.
+     *
+     * @param namespacedKey             The {@link NamespacedKey} to register this {@link AlliedAttackCheck} against.
+     * @param alliedAttackCheckFunction The {@link AlliedAttackCheck} to register.
+     */
+    public void registerAlliedAttackCheckFunction(@NotNull NamespacedKey namespacedKey, @NotNull AlliedAttackCheck alliedAttackCheckFunction) {
+        alliedAttackCheckFunctions.put(namespacedKey, alliedAttackCheckFunction);
+    }
+
+    /**
+     * Checks to see if the two provided {@link Entity entities} should be unable to attack each other, assuming {@link #areEntitiesAllied(Entity, Entity)} returns
+     * {@code true}. This first checks that {@link #areEntitiesAllied(Entity, Entity)} returns {@code true} before proceeding to check the matching {@link AlliedAttackCheck}.
+     * <p>
+     * This allows 3rd party plugins to anonymously register handling for their specific definition for when allies should be unable to attack each other.
+     * <p>
+     * The order of the two entities should not matter as well.
+     *
+     * @param entity1 The first {@link Entity} to check.
+     * @param entity2 The second {@link Entity} to check.
+     * @return {@code true} if the two {@link Entity entities} are considered allies by any registered {@link AlliedAttackCheck EntityAlliedFunctions}.
+     */
+    public Pair<Boolean, Optional<NamespacedKey>> shouldAlliesBeUnableToDamage(@NotNull Entity entity1, @NotNull Entity entity2) {
+        for (NamespacedKey namespacedKey : alliedAttackCheckFunctions.keySet()) {
+            AlliedAttackCheck alliedAttackCheckFunction = alliedAttackCheckFunctions.get(namespacedKey);
+            if (areEntitiesAllied(entity1, entity2, namespacedKey) && alliedAttackCheckFunction.shouldBeUnableToDamage(entity1, entity2)) {
+                return ImmutablePair.of(true, Optional.of(namespacedKey));
+            }
+        }
+        return ImmutablePair.of(false, Optional.empty());
     }
 }
