+++ src/main/java/us/eunoians/mcrpg/entity/check/EntityPetCheck.java
@@ -1,4 +1,4 @@
-package us.eunoians.mcrpg.ability.check;
+package us.eunoians.mcrpg.entity.check;
 
 import org.bukkit.entity.Entity;
 import org.jetbrains.annotations.NotNull;
@@ -6,5 +6,5 @@ import org.jetbrains.annotations.NotNull;
 //TODO javadoc (https://github.com/DiamondDagger590/McRPG/issues/180)
 public interface EntityPetCheck {
 
-    public boolean isEntityPetOf(@NotNull Entity entity1, @NotNull Entity entity2);
+    boolean isEntityPetOf(@NotNull Entity entity1, @NotNull Entity entity2);
 }
